requirements.txt
astroid==2.1.0
certifi==2019.3.9
chardet==3.0.4
defusedxml==0.5.0
django-allauth==0.50
django-tinymce4-lite==1.7.4
idna==2.8
isort==4.3.4
jsmin==2.2.2
lazy-object-proxy==1.3.1
mccabe==0.6.1
oauthlib==3.0.1
pep8==1.7.1
pylint==2.2.2
python3-openid==3.1.0
requests-oauthlib==1.2.0
six==1.12.0
typed-ast==1.5.2
urllib3==1.24.2
wrapt==1.11.0



cryptography==3.4.6
asgiref==3.3.4
autopep8==1.5.7
dj-database-url==0.5.0
Django==3.2.6
gunicorn==20.1.0
pycodestyle==2.7.0
python-decouple==3.4
sqlparse==0.4.1
toml==0.10.2
Unipath==1.1
whitenoise==5.3.0
wheel==0.37.0
psycopg2-binary>=2.8
django-cors-headers==3.7.0
drf-writable-nested==0.6.3
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
PyJWT==1.7.1
django-rq==2.4.0
channels-redis==3.2.0
rq~=1.7.0
channels~=3.0.3
pyfiglet==0.8.post1
django-crispy-forms==1.12.0

pytz==2021.1
django-tables2==2.4.0
Pillow==8.3.2
django-lifecycle==0.9.1
requests==2.26.0
websocket-client==1.2.1
django-debug-toolbar==3.2.2
django-rq==2.5.1
django-seed==0.3.1
